rename trimComments, to stripComments
