+++ src/lib/stripComments.ts
